FlxBox2D Destroy bodies and joints safely. Hopefully it won’t crash anymore.
Modified onScreen() for B2FlxShape and B2FlxJoint. It only checks if the AABB are within the screen. It doesn't look at the frameWidth and frameHeight.

Replace the default Box2DDebugRenderer with B2FlxDebug. The debug for shapes will be done at B2FlxShape. The joints and the contact are done in B2FlxDebug.

Bodies, joints, and debug work with FlxCamera. The camera can follow it. However it doesn’t work with offset. Also, it’s not tested with multiple cameras.

Chain shape has been implemented. It also supports loop and ghost edges.

Test demo
TestCart
It can be driven by A &D or on screen D-Pad. The follow camera has been set in this test.

Flx core
- MouseEvent, MOUSE_DOWN.
- FlxGame, dispatchEvent for MouseEvent.MOUSE_DOWN.
Not sure if this will become handy.